+++ src/app/_components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/app/_libs/utils";
 
 function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
