+++ src/app/_components/ui/switch.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as SwitchPrimitive from "@radix-ui/react-switch";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/app/_libs/utils";
 
 function Switch({
   className,
